Move <script> into <body> to pass validation
Failing validation causes strange errors through boto, which seems to
validate HTML against an external service; that external service returns
an error that includes a smart quote, which boto is then unable to
parse. The error that boto throws seems to indicate that you have a
smart quote in your HTML, which is rather hard to debug.